Remove AugmentedDataset from the API
