+++ pyriemann/preprocessing.py
@@ -200,43 +200,3 @@ class Whitening(BaseEstimator, TransformerMixin):
         """
         Xiw = self.inv_filters_.T @ X @ self.inv_filters_
         return Xiw
-
-
-class AugmentedDataset(BaseEstimator, TransformerMixin):
-    """This transformation allow to create an embedding version of the current dataset.
-    The implementation and the application is described in [1]_.
-
-    References
-    ----------
-    .. [1] Carrara, I., & Papadopoulo, T. (2023). Classification of BCI-EEG based on augmented covariance matrix.
-           arXiv preprint arXiv:2302.04508.
-           https://doi.org/10.48550/arXiv.2302.04508
-    """
-
-    def __init__(self, order=1, lag=1):
-        self.order = order
-        self.lag = lag
-
-    def fit(self, X, y):
-        return self
-
-    def transform(self, X):
-        if self.order == 1:
-            X_fin = X
-        else:
-            X_fin = []
-
-            for i in np.arange(X.shape[0]):
-                X_p = X[i][:, : -self.order * self.lag]
-                X_p = np.concatenate(
-                    [X_p]
-                    + [
-                        X[i][:, p * self.lag : -(self.order - p) * self.lag]
-                        for p in range(1, self.order)
-                    ],
-                    axis=0,
-                )
-                X_fin.append(X_p)
-            X_fin = np.array(X_fin)
-
-        return X_fin
